[forwarder] allow multiple endpoints
In order to do so, the config was slightly modified:
dd_url and api_key can now both receives a comma-separated "list", which is
transformed in a endpoints dict.
This one has the following format:
{
  'https://app.datadoghq.com': ['api_key_abc', 'api_key_def'],
  'https://app.example.com': ['api_key_xyz']
}

The forwarder is the only one using this new configuration.
It loops through all endpoints and sends the same payload.